update look of application WIP card component
