--- libs/service-portal/applications/src/screens/ApplicationListWip/mockApplicationList.ts
@@ -1,56 +1,0 @@
-import { Application, ApplicationTypes } from '@island.is/application/core'
-
-export const mockApplicationList: Application[] = [
-  {
-    id: '1',
-    state: 'in-progress',
-    applicant: '',
-    typeId: ApplicationTypes.EXAMPLE,
-    modified: new Date(),
-    created: new Date(0),
-    attachments: {},
-    answers: {},
-    externalData: {},
-    name: 'Umsókn um ökunám',
-    progress: 0.5,
-  },
-  {
-    id: '2',
-    state: 'in-progress',
-    applicant: '',
-    typeId: ApplicationTypes.EXAMPLE,
-    modified: new Date(),
-    created: new Date(0),
-    attachments: {},
-    answers: {},
-    externalData: {},
-    name: 'Umsókn um ökunám',
-    progress: 0.5,
-  },
-  {
-    id: '3',
-    state: 'in-progress',
-    applicant: '',
-    typeId: ApplicationTypes.EXAMPLE,
-    modified: new Date(),
-    created: new Date(0),
-    attachments: {},
-    answers: {},
-    externalData: {},
-    name: 'Umsókn um ökunám',
-    progress: 0.5,
-  },
-  {
-    id: '4',
-    state: 'in-progress',
-    applicant: '',
-    typeId: ApplicationTypes.EXAMPLE,
-    modified: new Date(),
-    created: new Date(0),
-    attachments: {},
-    answers: {},
-    externalData: {},
-    name: 'Umsókn um ökunám',
-    progress: 0.5,
-  },
-]
